Clean up and outsourcing. Updated colors. Added new progress bar.
